Move fake conduit to integration base
